+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/api.mustache
@@ -9,6 +9,9 @@ import javax.ws.rs.core.Response;
 import io.swagger.annotations.*;
 
 import java.util.List;
+{{#useBeanValidation}}
+import javax.validation.constraints.*;
+{{/useBeanValidation}}
 
 @Path("/{{baseName}}")
