+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/beanValidation.mustache
@@ -1,0 +1,53 @@
+{{#required}}
+  @NotNull
+{{/required}}
+{{#pattern}}
+  @Pattern(regexp="{{pattern}}")
+{{/pattern}}
+{{#minLength}}
+{{#maxLength}}
+  @Size(min={{minLength}},max={{maxLength}})
+{{/maxLength}}
+{{/minLength}}
+{{#minLength}}
+{{^maxLength}}
+  @Size(min={{minLength}})
+{{/maxLength}}
+{{/minLength}}
+{{^minLength}}
+{{#maxLength}}
+  @Size(max={{maxLength}})
+  {{/maxLength}}
+  {{/minLength}}
+{{#minItems}}
+{{#maxItems}}
+  @Size(min={{minItems}},max={{maxItems}})
+{{/maxItems}}
+{{/minItems}}
+{{#minItems}}
+{{^maxItems}}
+  @Size(min={{minItems}})
+{{/maxItems}}
+{{/minItems}}
+{{^minItems}}
+{{#maxItems}}
+  @Size(max={{maxItems}})
+{{/maxItems}}
+{{/minItems}}
+{{! check for integer / number=decimal type}}
+{{#isInteger}}
+{{#minimum}}
+  @Min({{minimum}})
+{{/minimum}}
+{{#maximum}}
+  @Max({{maximum}})
+{{/maximum}}
+{{/isInteger}}
+{{^isInteger}}
+{{#minimum}}
+  @DecimalMin("{{minimum}}")
+{{/minimum}}
+{{#maximum}}
+  @DecimalMax("{{maximum}}")
+{{/maximum}}
+{{/isInteger}}
