+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/enumClass.mustache
@@ -1,16 +1,31 @@
-import javax.xml.bind.annotation.XmlEnum;
-import javax.xml.bind.annotation.XmlType;
+public enum {{datatypeWithEnum}} {
 
-@XmlType(name="{{classname}}")
-@XmlEnum
-public enum {{classname}} {
-    {{#allowableValues}}{{.}}{{^-last}}, {{/-last}}{{#-last}};{{/-last}}{{/allowableValues}}
-    
-    public String value() {
-        return name();
+    {{#allowableValues}}
+    {{#enumVars}}{{name}}({{datatype}}.valueOf({{{value}}})){{^-last}}, {{/-last}}{{#-last}};{{/-last}}{{/enumVars}}
+    {{/allowableValues}}
+
+
+    private {{datatype}} value;
+
+    {{datatypeWithEnum}} ({{datatype}} v) {
+        value = v;
     }
 
-    public static {{classname}} fromValue(String v) {
-        return valueOf(v);
+    public String value() {
+        return value;
+    }
+
+    @Override
+    public String toString() {
+        return String.valueOf(value);
+    }
+
+    public static {{datatypeWithEnum}} fromValue(String v) {
+        for ({{#datatypeWithEnum}}{{{.}}}{{/datatypeWithEnum}}{{^datatypeWithEnum}}{{{classname}}}{{/datatypeWithEnum}} b : {{#datatypeWithEnum}}{{{.}}}{{/datatypeWithEnum}}{{^datatypeWithEnum}}{{{classname}}}{{/datatypeWithEnum}}.values()) {
+            if (String.valueOf(b.value).equals(v)) {
+                return b;
+            }
+        }
+        return null;
     }
 }
