+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/generatedAnnotation.mustache
@@ -1,1 +1,3 @@
+{{^hideGenerationTimestamp}}
 @javax.annotation.Generated(value = "{{generatorClass}}", date = "{{generatedDate}}")
+{{/hideGenerationTimestamp}}
