+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/model.mustache
@@ -2,13 +2,18 @@ package {{package}};
 
 {{#imports}}import {{import}};
 {{/imports}}
+{{#useBeanValidation}}
+import javax.validation.constraints.*;
+{{/useBeanValidation}}
 
 {{#models}}
 {{#model}}{{#description}}
 /**
  * {{description}}
  **/{{/description}}
-{{#isEnum}}{{>enumClass}}{{/isEnum}}
+{{#isEnum}}
+{{>enumOuterClass}}
+{{/isEnum}}
 {{^isEnum}}{{>pojo}}{{/isEnum}}
 {{/model}}
 {{/models}}
