+++ modules/swagger-codegen/src/main/resources/JavaJaxRS/spec/pom.mustache
@@ -69,6 +69,15 @@
         </exclusion>
       </exclusions>
     </dependency>
+    {{#useBeanValidation}}
+    <!-- Bean Validation API support -->
+    <dependency>
+        <groupId>javax.validation</groupId>
+        <artifactId>validation-api</artifactId>
+        <version>1.1.0.Final</version>
+        <scope>provided</scope>
+    </dependency>
+{{/useBeanValidation}}
   </dependencies>
   <properties>
     <junit-version>4.8.1</junit-version>
