+++ samples/server/petstore/jaxrs-spec/pom.xml
@@ -1,9 +1,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>wasdev</groupId>
-  <artifactId>autogen-server</artifactId>
+  <groupId>io.swagger</groupId>
+  <artifactId>swagger-jaxrs-server</artifactId>
   <packaging>war</packaging>
-  <name>autogen-server</name>
+  <name>swagger-jaxrs-server</name>
   <version>1.0.0</version>
   <build>
     <sourceDirectory>src/main/java</sourceDirectory>
@@ -68,6 +68,13 @@
           <groupId>org.beanshell</groupId>
         </exclusion>
       </exclusions>
+    </dependency>
+    <!-- Bean Validation API support -->
+    <dependency>
+        <groupId>javax.validation</groupId>
+        <artifactId>validation-api</artifactId>
+        <version>1.1.0.Final</version>
+        <scope>provided</scope>
     </dependency>
   </dependencies>
   <properties>
