+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/api/RestApplication.java
@@ -1,4 +1,4 @@
-package com.ibm.ws.petstoresample;
+package io.swagger.api;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
