+++ samples/server/petstore/jaxrs-spec/src/gen/java/io/swagger/model/ModelApiResponse.java
@@ -1,0 +1,110 @@
+package io.swagger.model;
+
+import io.swagger.annotations.ApiModel;
+import javax.validation.constraints.*;
+
+
+/**
+ * Describes the result of uploading an image resource
+ **/
+import io.swagger.annotations.*;
+import java.util.Objects;
+@ApiModel(description = "Describes the result of uploading an image resource")
+
+public class ModelApiResponse   {
+  
+  private Integer code = null;
+  private String type = null;
+  private String message = null;
+
+  /**
+   **/
+  public ModelApiResponse code(Integer code) {
+    this.code = code;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "null", value = "")
+  public Integer getCode() {
+    return code;
+  }
+  public void setCode(Integer code) {
+    this.code = code;
+  }
+
+  /**
+   **/
+  public ModelApiResponse type(String type) {
+    this.type = type;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "null", value = "")
+  public String getType() {
+    return type;
+  }
+  public void setType(String type) {
+    this.type = type;
+  }
+
+  /**
+   **/
+  public ModelApiResponse message(String message) {
+    this.message = message;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "null", value = "")
+  public String getMessage() {
+    return message;
+  }
+  public void setMessage(String message) {
+    this.message = message;
+  }
+
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    ModelApiResponse _apiResponse = (ModelApiResponse) o;
+    return Objects.equals(code, _apiResponse.code) &&
+        Objects.equals(type, _apiResponse.type) &&
+        Objects.equals(message, _apiResponse.message);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(code, type, message);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class ModelApiResponse {\n");
+    
+    sb.append("    code: ").append(toIndentedString(code)).append("\n");
+    sb.append("    type: ").append(toIndentedString(type)).append("\n");
+    sb.append("    message: ").append(toIndentedString(message)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
+}
