--- samples/server/petstore/jaxrs-spec/src/main/java/com/ibm/ws/petstoresample/model/Order.java
@@ -1,164 +1,0 @@
-package com.ibm.ws.petstoresample.model;
-
-
-
-
-import io.swagger.annotations.*;
-import java.util.Objects;
-
-
-public class Order   {
-  
-  private Long id = null;
-  private Long petId = null;
-  private Integer quantity = null;
-  private javax.xml.datatype.XMLGregorianCalendar shipDate = null;
-  private String status = null;
-  private Boolean complete = null;
-
-  /**
-   **/
-  public Order id(Long id) {
-    this.id = id;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Long getId() {
-    return id;
-  }
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  /**
-   **/
-  public Order petId(Long petId) {
-    this.petId = petId;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Long getPetId() {
-    return petId;
-  }
-  public void setPetId(Long petId) {
-    this.petId = petId;
-  }
-
-  /**
-   **/
-  public Order quantity(Integer quantity) {
-    this.quantity = quantity;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Integer getQuantity() {
-    return quantity;
-  }
-  public void setQuantity(Integer quantity) {
-    this.quantity = quantity;
-  }
-
-  /**
-   **/
-  public Order shipDate(javax.xml.datatype.XMLGregorianCalendar shipDate) {
-    this.shipDate = shipDate;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public javax.xml.datatype.XMLGregorianCalendar getShipDate() {
-    return shipDate;
-  }
-  public void setShipDate(javax.xml.datatype.XMLGregorianCalendar shipDate) {
-    this.shipDate = shipDate;
-  }
-
-  /**
-   * Order Status
-   **/
-  public Order status(String status) {
-    this.status = status;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "Order Status")
-  public String getStatus() {
-    return status;
-  }
-  public void setStatus(String status) {
-    this.status = status;
-  }
-
-  /**
-   **/
-  public Order complete(Boolean complete) {
-    this.complete = complete;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Boolean getComplete() {
-    return complete;
-  }
-  public void setComplete(Boolean complete) {
-    this.complete = complete;
-  }
-
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Order order = (Order) o;
-    return Objects.equals(id, order.id) &&
-        Objects.equals(petId, order.petId) &&
-        Objects.equals(quantity, order.quantity) &&
-        Objects.equals(shipDate, order.shipDate) &&
-        Objects.equals(status, order.status) &&
-        Objects.equals(complete, order.complete);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, petId, quantity, shipDate, status, complete);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Order {\n");
-    
-    sb.append("    id: ").append(toIndentedString(id)).append("\n");
-    sb.append("    petId: ").append(toIndentedString(petId)).append("\n");
-    sb.append("    quantity: ").append(toIndentedString(quantity)).append("\n");
-    sb.append("    shipDate: ").append(toIndentedString(shipDate)).append("\n");
-    sb.append("    status: ").append(toIndentedString(status)).append("\n");
-    sb.append("    complete: ").append(toIndentedString(complete)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
