--- samples/server/petstore/jaxrs-spec/src/main/java/com/ibm/ws/petstoresample/model/Pet.java
@@ -1,168 +1,0 @@
-package com.ibm.ws.petstoresample.model;
-
-import com.ibm.ws.petstoresample.model.Category;
-import com.ibm.ws.petstoresample.model.Tag;
-import java.util.ArrayList;
-import java.util.List;
-
-
-
-import io.swagger.annotations.*;
-import java.util.Objects;
-
-
-public class Pet   {
-  
-  private Long id = null;
-  private Category category = null;
-  private String name = null;
-  private List<String> photoUrls = new ArrayList<String>();
-  private List<Tag> tags = new ArrayList<Tag>();
-  private String status = null;
-
-  /**
-   **/
-  public Pet id(Long id) {
-    this.id = id;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Long getId() {
-    return id;
-  }
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  /**
-   **/
-  public Pet category(Category category) {
-    this.category = category;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public Category getCategory() {
-    return category;
-  }
-  public void setCategory(Category category) {
-    this.category = category;
-  }
-
-  /**
-   **/
-  public Pet name(String name) {
-    this.name = name;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "doggie", required = true, value = "")
-  public String getName() {
-    return name;
-  }
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  /**
-   **/
-  public Pet photoUrls(List<String> photoUrls) {
-    this.photoUrls = photoUrls;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", required = true, value = "")
-  public List<String> getPhotoUrls() {
-    return photoUrls;
-  }
-  public void setPhotoUrls(List<String> photoUrls) {
-    this.photoUrls = photoUrls;
-  }
-
-  /**
-   **/
-  public Pet tags(List<Tag> tags) {
-    this.tags = tags;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "")
-  public List<Tag> getTags() {
-    return tags;
-  }
-  public void setTags(List<Tag> tags) {
-    this.tags = tags;
-  }
-
-  /**
-   * pet status in the store
-   **/
-  public Pet status(String status) {
-    this.status = status;
-    return this;
-  }
-
-  
-  @ApiModelProperty(example = "null", value = "pet status in the store")
-  public String getStatus() {
-    return status;
-  }
-  public void setStatus(String status) {
-    this.status = status;
-  }
-
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Pet pet = (Pet) o;
-    return Objects.equals(id, pet.id) &&
-        Objects.equals(category, pet.category) &&
-        Objects.equals(name, pet.name) &&
-        Objects.equals(photoUrls, pet.photoUrls) &&
-        Objects.equals(tags, pet.tags) &&
-        Objects.equals(status, pet.status);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, category, name, photoUrls, tags, status);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Pet {\n");
-    
-    sb.append("    id: ").append(toIndentedString(id)).append("\n");
-    sb.append("    category: ").append(toIndentedString(category)).append("\n");
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    photoUrls: ").append(toIndentedString(photoUrls)).append("\n");
-    sb.append("    tags: ").append(toIndentedString(tags)).append("\n");
-    sb.append("    status: ").append(toIndentedString(status)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
